Use fake revision if none given
When using this to test sourcemap behavior, you want to see a
different revision uploaded each time, but you don't actually care
what the revision name _is_, and having to set it manually to a
different value in each run is mildly inconvenient.

If no revision value is given, generate something that looks like
a Git commit SHA with a "-fake" prefix, and use that as a revision.