Allow granular configuration beyond frameworks
This allows us to define different configurations for different
subsets of a framework's applications (`react/cra` and `react/vite`)
or potentially overriding the configuration for one specific version.

It also explicitly specifies which packages to link, rather than
relying on the framework folder name matching the name of an
`@appsignal/` package.

Finally, it changes the clean command so that it removes the specific
folder that the app uses for the build, as described in the config
for that app, instead of always removing the `build` and `dist`
folders.